Moving once per app launch to Application class. Adding Custom Notification manager
